Initial and finished project
Created a webservice using gin framework with one GET and one POST method. The POST method will take in a JSON, bind that JSON to a Receipt object, generate an id, save the Receipt object to a map called receiptDatabase by id, and return a JSON containing an id for the receipt. The POST method will also calculate the total number of points awarded and store that information in a map called pointsDatabase by id. The GET method will access the database by id and return a Points JSON object.